Refactoring entry file
