+++ .gitignore
@@ -19,6 +19,7 @@ yarn.lock
 *.sln
 coverage
 *.orig
+*.log
 config/secrets.env.js
 
 cypress/videos
