--- chromedriver.log
@@ -1,6 +1,0 @@
-Starting ChromeDriver 2.43.600229 (3fae4d0cda5334b4f533bede5a4787f7b832d052) on port 9515
-Only local connections are allowed.
-IOVARendererID property not found 
-IOVARendererID property not found 
-IOVARendererID property not found 
-IOVARendererID property not found
