+++ package.json
@@ -1,9 +1,10 @@
 {
-  "name": "vue-quick-dash",
+  "name": "@unicorn/quick-dash",
   "version": "1.0.0",
-  "description": "A Vue.js project",
+  "description": "Quick Dashboards",
   "author": "Darryn Ten, Steve Ops",
   "private": true,
+  "module": "src/main.js",
   "scripts": {
     "dev": "node build/dev-server.js",
     "start": "npm run dev",
@@ -18,32 +19,32 @@
     "axios": "^0.18.0",
     "unicorn-vue-datatable": "^1.2.1",
     "unicorn-vue-toaster": "^1.0.1",
-    "vee-validate": "^2.1.0-beta.11",
+    "vee-validate": "^2.1.3",
     "vue": "^2.5.17",
     "vue-avatar-or-initials": "^1.0.7",
     "vue-axios": "^2.1.4",
-    "vue-router": "^3.0.1"
+    "vue-router": "^3.0.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.1.6",
-    "@babel/plugin-proposal-class-properties": "^7.0.0",
-    "@babel/plugin-proposal-decorators": "^7.1.6",
-    "@babel/plugin-proposal-export-namespace-from": "^7.0.0",
-    "@babel/plugin-proposal-function-sent": "^7.0.0",
-    "@babel/plugin-proposal-json-strings": "^7.0.0",
-    "@babel/plugin-proposal-numeric-separator": "^7.0.0",
-    "@babel/plugin-proposal-throw-expressions": "^7.0.0",
-    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
-    "@babel/plugin-syntax-import-meta": "^7.0.0",
-    "@babel/plugin-transform-runtime": "^7.0.0",
+    "@babel/core": "^7.2.0",
+    "@babel/plugin-proposal-class-properties": "^7.2.0",
+    "@babel/plugin-proposal-decorators": "^7.2.0",
+    "@babel/plugin-proposal-export-namespace-from": "^7.2.0",
+    "@babel/plugin-proposal-function-sent": "^7.2.0",
+    "@babel/plugin-proposal-json-strings": "^7.2.0",
+    "@babel/plugin-proposal-numeric-separator": "^7.2.0",
+    "@babel/plugin-proposal-throw-expressions": "^7.2.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@babel/plugin-syntax-import-meta": "^7.2.0",
+    "@babel/plugin-transform-runtime": "^7.2.0",
     "@babel/polyfill": "^7.0.0",
     "@babel/preset-env": "^7.1.6",
     "@babel/register": "^7.0.0",
-    "@babel/runtime": "^7.1.5",
+    "@babel/runtime": "^7.2.0",
     "@vue/test-utils": "1.0.0-beta.25",
     "@websanova/vue-auth": "^2.21.14-beta",
-    "ajv": "^6.5.5",
-    "autoprefixer": "^9.3.1",
+    "ajv": "^6.6.1",
+    "autoprefixer": "^9.4.0",
     "babel-eslint": "^10.0.1",
     "babel-loader": "^8.0.4",
     "babel-plugin-istanbul": "^5.1.0",
@@ -56,16 +57,16 @@
     "copy-webpack-plugin": "^4.6.0",
     "cross-env": "^5.2.0",
     "cross-spawn": "^6.0.5",
-    "cypress": "^3.1.1",
     "css-loader": "^1.0.1",
+    "cypress": "^3.1.3",
     "eslint": "^5.9.0",
     "eslint-config-standard": "^12.0.0",
     "eslint-friendly-formatter": "^4.0.1",
     "eslint-loader": "^2.1.1",
     "eslint-plugin-chai-friendly": "^0.4.1",
-    "eslint-plugin-html": "^4.0.6",
+    "eslint-plugin-html": "^5.0.0",
     "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-node": "^7.0.1",
+    "eslint-plugin-node": "^8.0.0",
     "eslint-plugin-promise": "^4.0.1",
     "eslint-plugin-standard": "^4.0.0",
     "eventsource-polyfill": "^0.9.6",
@@ -74,7 +75,7 @@
     "flush-promises": "^1.0.2",
     "friendly-errors-webpack-plugin": "^1.7.0",
     "html-webpack-plugin": "^3.2.0",
-    "http-proxy-middleware": "^0.19.0",
+    "http-proxy-middleware": "^0.19.1",
     "inject-loader": "^4.0.1",
     "karma": "^3.1.1",
     "karma-chrome-launcher": "^2.2.0",
@@ -89,14 +90,14 @@
     "opn": "^5.4.0",
     "optimize-css-assets-webpack-plugin": "^5.0.1",
     "ora": "^3.0.0",
-    "portfinder": "^1.0.19",
+    "portfinder": "^1.0.20",
     "rimraf": "^2.6.0",
     "sass-loader": "^7.1.0",
-    "sass-resources-loader": "^1.3.4",
+    "sass-resources-loader": "^2.0.0",
     "semver": "^5.6.0",
     "shelljs": "^0.8.3",
     "sinon": "^7.1.1",
-    "sinon-chai": "^3.2.0",
+    "sinon-chai": "^3.3.0",
     "uglifyjs-webpack-plugin": "^1.3.0",
     "url-loader": "^1.1.2",
     "vue-loader": "^15.4.2",
@@ -105,7 +106,7 @@
     "vue-svg-loader": "^0.10.0",
     "vue-template-compiler": "^2.5.17",
     "vuex": "^3.0.1",
-    "webpack": "^4.26.0",
+    "webpack": "^4.26.1",
     "webpack-bundle-analyzer": "^3.0.3",
     "webpack-cli": "^3.1.2",
     "webpack-dev-middleware": "^3.4.0",
@@ -113,8 +114,8 @@
     "webpack-merge": "^4.1.4"
   },
   "engines": {
-    "node": ">= 7.0.0",
-    "npm": ">= 3.0.0"
+    "node": ">= 10.0.0",
+    "npm": ">= 6.0.0"
   },
   "browserslist": [
     "> 1%",
