+++ src/api/user.js
@@ -1,5 +1,5 @@
 import Vue from 'vue'
-import store from '@/store'
+import store from '../store'
 
 export function loadUserDetails(userId) {
   return Vue.axios.get('api/me')
