+++ src/assets/sass/base.scss
@@ -1,4 +1,4 @@
-@import "./variables";
+@import "./_variables.scss";
 
 /**
 Enable the specific ones you want to use, do not enable all!
