+++ src/comms.js
@@ -1,0 +1,12 @@
+import Vue from 'vue'
+import axios from 'axios'
+import VueAxios from 'vue-axios'
+
+export default {
+  init: () => {
+    Vue.use(VueAxios, axios)
+    Vue.axios.defaults.baseURL = process.env.apiUrl
+    Vue.axios.defaults.headers.post['Content-Type'] = 'application/json'
+    // Vue.axios.defaults.headers.common['Debug-Token'] = process.env.debugToken
+  }
+}
