+++ src/router/UserRoutes.js
@@ -6,17 +6,17 @@ export default [
     meta: {
       main: true,
       label: 'Home',
-      icon: require('@/components/SideBar/icons/home.svg')
+      icon: require('@/components/SideBar/icons/home.svg').default
     }
   },
   {
     name: 'Account',
     path: '/account',
-    component: require('@/pages/User/Details').default,
+    component: require('@/pages/User/Details.vue').default,
     meta: {
       main: true,
       label: 'Account',
-      icon: require('@/components/SideBar/icons/PersonalAccount.svg')
+      icon: require('@/components/SideBar/icons/PersonalAccount.svg').default
     }
   }
 ]
