+++ src/router/index.js
@@ -1,7 +1,18 @@
+import Vue from 'vue'
 import Router from 'vue-router'
 import SharedRoutes from './SharedRoutes'
+import store from '@/store'
 
-export default new Router({
+let router = new Router({
   mode: 'history',
   routes: [...SharedRoutes]
 })
+
+Vue.use(Router)
+Vue.router = router
+
+Vue.router.afterEach(function (to, from) {
+  store.commit('showSideBar', false)
+})
+
+export default router
